refactor: simplify HeroSection component and enhance user experience
- Removed unused imports and state variables to streamline the HeroSection component.
- Integrated new UI components for URL input and loading overlay, improving the overall user interface.
- Updated analysis summary handling to utilize a more structured approach with the new MockResult type.
- Enhanced loading indicators and user feedback during analysis processes.
- Improved code readability and maintainability by reducing complexity in state management and rendering logic.
